cadastro instituicao

+++ backend/UcsHubAPI.Service/Services/ProjectService.cs
@@ -71,7 +71,7 @@ namespace UcsHubAPI.Service.Services
             }
 
             InstitutionRepository InstitutionRepository = new InstitutionRepository(_appSettings.ConnString);
-            project.Institution = InstitutionRepository.GetById(project.Institution.Id);
+            project.Institution = InstitutionRepository.GetById((int)project.Institution.Id);
 
 
             ProductionRepository productionRepository = new ProductionRepository(_appSettings.ConnString);
